paracon_dashboard_app_upload_mode/requirements.txt
streamlit>=1.36
pandas>=2.2
openpyxl>=3.1
plotly>=5.24
python-dotenv>=1.0
requests>=2.31
